refactoring the enum simd due to tests (errors, float only) on the BG/Q errors ([HPCTM-73])
In the previous version of cyme the enum simd encapuslates the technology and the
size of the SIMD register. Using this enum I was able to make specialization on my
final wrappersr, and calculate some meta-function to calculate offset for padding.

The main hypothesis was: divide the size of the SIMD technology ( e.g. 32 [Byte] for avx)
by the size of value_type. The hypothese works every where except on BG/Q where
the QPX can only perform operations on 4 float and not 8.

Consequently I separate this enum to the size of the register. I have now a trait class
which associates value of the SIMD register and alignement to the needed SIMD technology.

Change-Id: I89bb299a7ae3a42a2803b2bb140f7d2244cae1ff