optimization negate operator
This optimization is a first one of series done on my free time.

During "free" friday and discussion with Sam, we realize
it will be possible to make optimization directly on the tree.
Especially to capture FMA.

This first optimization concerns the negation, specially the
operations -(-(a[0])) (thank you, Francesco for the idead)
which now remplace a[0] (do nothing),  thus I avoid two sucessives
negate operations

I add a variation of the operators negate and modify the class-> struct
for the vec_class, as I need an access to the original data.
I add the corresponding tests.

The next optimization corresponds to the profiler which is +- ready,
but it necessiates at least c++11.

Introduce the solver pow for cyme

the algorithm (optimized) is the following:

x^n = (x^n/2)^2 n even
x^n = n(x^n/2)^2 n odd

tests and documentation are updated

Change-Id: I5ef24f2da351e15a6eefa76948bb8ab1e8ea71c9