add log for cyme
details: we introduce the log for cyme with the following algo:
 x = s*x0*2^n (where s the sign, always +), x0 (fraction) belongs between 1<=x0<2 and n the exponent (integer).
        Thus, log2(x) = log2(x0) + log2(2^n)
              log2(x) = log2(x0) + n
              log(x) = log(2) * log2(x)
        x0 and n are determinated by bit tips (manipulation)

The algo utilized the same tips with pade approximation and Heuler method.
For sumplicity the Euler method is now generic (math/detail/remez.hpp) it depends
from a template parameter coefficient which given in the callee.

Nevertheless it looks like the precision is not optimum for the float intel (I already increase the error for the test)
specially for number bewteen [0,1] (more close to one). The CEPHES library make distinction log(x) and log(x+1)
to investigate

Change-Id: I56d071736ba484b6bce9a19a58ff908a0fb692c4